requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
tavily-python>=0.5.0
PyMuPDF>=1.25.0
arxiv>=2.1.0
python-dotenv>=1.0.0
sse-starlette>=2.0.0
httpx>=0.28.0
tenacity>=9.0.0
slowapi>=0.1.9
pytest>=8.0.0
pytest-asyncio>=0.24.0
